Atualização para o Laravel 12

+++ config/filesystems.php
@@ -51,7 +51,7 @@ return [
         'public' => [
             'driver' => 'local',
             'root' => storage_path('app/public'),
-            'url' => env('APP_URL').'/storage',
+            'url' => env('APP_URL') . '/storage',
             'visibility' => 'public',
         ],
 
@@ -60,7 +60,7 @@ return [
             'host' => env('FTP_HOST'),
             'username' => env('FTP_USERNAME'),
             'password' => env('FTP_PASSWORD'),
-            'port' => env('FTP_PORT', 21),
+            'port' => (int)env('FTP_PORT', 21),
             'root' => env('FTP_ROOT'),
         ],
